proper projectile to unit damage, unit voxelmaps, deaths and shields
- units now have proper voxelmaps
- units receive damage properly from projectiles
- units emit proper sounds when hit
- units die from wounds
- units drop items when dead
- items other than weapons now display ammo/durability indicator

+++ game/state/city/vehiclemission.cpp
@@ -1072,8 +1072,8 @@ void VehicleMission::start(GameState &state, Vehicle &v)
 									xPos += vec.x;
 									yPos += vec.y;
 								}
-								xPos /= rooftop.size();
-								yPos /= rooftop.size();
+								xPos /= (int)rooftop.size();
+								yPos /= (int)rooftop.size();
 								// ensure object is within roof
 								if (rooftop.find({xPos, yPos}) == rooftop.end())
 								{
